update loading txns to grab latest by datetime string; add script to grab new txns (daily, ideally)
